requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge-score
nltk    
py7zr
pandas
tqdm
matplotlib
torch
notebook
boto3
mypy-boto3-s3
evaluate
pyYAML
python-box==7.3.2
ensure==1.0.2
fastapi==0.95.2
uvicorn==0.21.1
jinja2==3.1.2
-e .
